added user auth for users cart_by_id to proctect users info

+++ routes/users/cart/cart_by_id.py
@@ -1,16 +1,18 @@
 from flask_restful import Resource
 from config import api
 from models.models import Cart
-from flask import request
+from flask import request, session
 from marshmallow_schemas.cart import cart_schema
 
 class CartByID(Resource):
   def patch(self, id):
-    json = request.get_json()
     cart = Cart.query.filter_by(id=id).first()
     if not cart:
       return {"error": "Cart does not exist"}, 400
-    
+    if cart.user_id != session.get('user_id'):
+      return {"error": "Unauthorized"}, 403
+
+    json = request.get_json()
     cart.update_db(json)
     return cart_schema.dump(cart), 200
 
@@ -18,7 +20,9 @@ class CartByID(Resource):
     cart = Cart.query.filter_by(id=id).first()
     if not cart:
       return {"error": "Cart does not exist"}, 400
-    
+    if cart.user_id != session.get('user_id'):
+      return {"error": "Unauthorized"}, 403
+
     cart.delete_db()
     return {}, 200
